Complete 4.21-4.22

+++ part4/bloglist/controllers/blogs.js
@@ -18,14 +18,12 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-    const body = request.body
-    const token = request.token
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-    if (!token || !decodedToken.id) {
+    const user = request.user
+    if (!user) {
         return response.status(401).json({ error: 'token missing or invalid' })
     }
-    const user = await User.findById(decodedToken.id)
 
+    const body=request.body
     const blog = new Blog({
         title: body.title,
         author: body.author || 'unknown',
@@ -41,8 +39,19 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+    if (!blog)
+        return response.status(204).end()
+
+    const user = request.user
+    if (!user) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    if (blog.user.toString() === user._id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        return response.status(204).end()
+    }
+    return response.status(401).json({ error: 'cannon remove blogs that belong to someone else' })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
